Add patchy fix
Add a rather patchy (and not generally correct) solution to a problem. Previously, subnormal numbers were not returned
correctly, due to the return code expecting all mantissas to have a leading implicit bit that may be removed. This fixes
that issue, but introduces a very subtle new one. As currently implemented, the mantissa of a subnormal float will be
rounded to 24 bits of precision, rather than the correct 23. This results in subnormal floats essentially being
"double-rounded", causing small deviations in a very small number of edge cases that still must be addressed.

A possible fix takes the form of shifting subnormal mantissas one bit to the right after the addition routine and before
the rounding routine, so that they are correctly rounded to 23 bits of precision, and the return code may revert to its
previous form.